requirements.txt
astroid==2.2.2
backcall==0.1.0
backports.csv==1.0.7
beautifulsoup4==4.7.1
certifi==2018.11.29
cffi==1.12.2
cfscrape==1.9.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
docopt==0.6.2
et-xmlfile==1.0.1
Flask==1.0.2
future==0.17.1
gevent==1.4.0
greenlet==0.4.15
gspread==3.1.0
gunicorn==19.9.0
hdfs==2.2.2
html5lib==1.0.1
httplib2==0.12.1
idna==2.8
ipython==7.3.0
ipython-genutils==0.2.0
isort==4.3.10
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.3
jieba==0.39
Jinja2==2.10.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
line-bot-sdk==1.8.0
lxml==4.3.2
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
numpy==1.16.2
oauth2client==4.1.3
odfpy==1.4.0
openpyxl==2.6.1
pandas==0.24.1
parso==0.3.4
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
Pygments==2.3.1
pylint==2.3.1
pylint-flask==0.6
pylint-plugin-utils==0.5
PyMySQL==0.9.3
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==4.0
scikit-learn==0.20.3
scipy==1.2.1
selenium==3.141.0
six==1.12.0
soupsieve==1.8
tablib==0.13.0
traitlets==4.3.2
typed-ast==1.3.1
urllib3==1.24.1
virtualenv==16.4.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
